# docs/api/openapi-generated.yaml
components:
  schemas:
...
    ExternalTemplateDataResponse:
      description: External template data response.
      properties:
        data:
          items: {}
          type: array
        projectId:
          format: uuid
          type: string
        templateId:
          format: uuid
          type: string
      required:
      - templateId
      - projectId
      - data
      type: object
...
    ProblemDetails:
      allOf:
      - description: RFC 7807 extension fields (e.g. for 409 `StaleConversation` extra data).
        type:
        - object
        - 'null'
      - properties:
          detail:
            description: Detailed error description.
            type:
            - string
            - 'null'
          errors:
            description: Field-level validation errors.
            items:
              $ref: '#/components/schemas/ValidationError'
            type:
            - array
            - 'null'
          instance:
            description: URI of the specific resource where the error occurred.
            type:
            - string
            - 'null'
          status:
            description: HTTP status code.
            format: int32
            minimum: 0
            type: integer
          title:
            description: Human-readable error summary.
            type: string
          type:
            description: Error type URI for programmatic identification.
            type: string
        required:
        - type
        - title
        - status
        type: object
      description: RFC 7807 Problem Details error response.
...
    ValidationError:
      description: A single field validation error.
      properties:
        code:
          description: Error code for programmatic handling (e.g. "required", "`too_long`").
          type: string
        field:
          description: Field path (e.g. "name", "config.apiKey").
          type: string
        message:
          description: Human-readable error message.
          type: string
      required:
      - field
      - message
      - code
      type: object
...
  securitySchemes:
    bearer_auth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: AI-assisted conference/event project management system
  license:
    name: ''
  title: Conf-Ops API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /external/v1/projects/{projectId}/task-templates:
    get:
      description: |-
        Requires API Key authentication via X-API-Key header.

        # Errors

        Returns `ProblemDetails` on authorization failure.
      operationId: list_external_templates
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          description: Template list
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: List task templates (external API).
      tags:
      - external-data
  /external/v1/projects/{projectId}/task-templates/{templateId}/data:
    get:
      description: |-
        Requires API Key authentication via X-API-Key header.

        # Errors

        Returns `ProblemDetails` on not found or authorization failure.
      operationId: get_template_data
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalTemplateDataResponse'
          description: Template aggregated data
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: Get aggregated data for a task template (external API).
      tags:
      - external-data
  /external/v1/projects/{projectId}/tasks/{taskId}/data:
    get:
      description: |-
        # Errors

        Returns `ProblemDetails` on authorization failure.
      operationId: get_task_data
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          description: Task data
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: Get task data (external API).
      tags:
      - external-data
    put:
      description: |-
        # Errors

        Returns `ProblemDetails` on authorization failure.
      operationId: update_task_data
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          description: Task data updated
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: Update task data (external API).
      tags:
      - external-data
...